Added files via upload
Here are all the files from my project folder that I used to create the user interface.  This will be saved here for saving and editing the code.  Feel free to edit any piece of code in any of the .h or .cpp files.  I'll be sure to document each ui method after upload